Change Version Handling
Paper stated that parsing the plugin version through the craftbukkit package won't be supported for much longer. Therefor we now use the long-standing API for the BukkitVersion to check which refection provider to use. During testing we also found a couple of issues where on some versions of spigot the NetworkManager field is private. This is also fixed in this commit.